Cleans up useEffect hooks
